PLAT-10424 pass custom BotAuthenticator to SymphonyBDK Constructor
