+++ symphony-bdk-core/src/test/java/com/symphony/bdk/core/auth/impl/AuthenticatorFactoryImplTest.java
@@ -1,4 +1,4 @@
-package com.symphony.bdk.core.auth;
+package com.symphony.bdk.core.auth.impl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -6,13 +6,11 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import com.symphony.bdk.core.auth.BotAuthenticator;
+import com.symphony.bdk.core.auth.ExtensionAppAuthenticator;
+import com.symphony.bdk.core.auth.AuthenticatorFactory;
+import com.symphony.bdk.core.auth.OboAuthenticator;
 import com.symphony.bdk.core.auth.exception.AuthInitializationException;
-import com.symphony.bdk.core.auth.impl.BotAuthenticatorCertImpl;
-import com.symphony.bdk.core.auth.impl.BotAuthenticatorRsaImpl;
-import com.symphony.bdk.core.auth.impl.ExtensionAppAuthenticatorCertImpl;
-import com.symphony.bdk.core.auth.impl.ExtensionAppAuthenticatorRsaImpl;
-import com.symphony.bdk.core.auth.impl.OboAuthenticatorCertImpl;
-import com.symphony.bdk.core.auth.impl.OboAuthenticatorRsaImpl;
 import com.symphony.bdk.core.client.ApiClientFactory;
 import com.symphony.bdk.core.config.model.BdkCertificateConfig;
 import com.symphony.bdk.core.config.model.BdkConfig;
@@ -33,9 +31,9 @@ import java.util.UUID;
 import java.util.function.Supplier;
 
 /**
- * Test class for the {@link AuthenticatorFactory}.
+ * Test class for the {@link AuthenticatorFactoryImpl}.
  */
-class AuthenticatorFactoryTest {
+class AuthenticatorFactoryImplTest {
 
   private static final ApiClient DUMMY_API_CLIENT = new ApiClientBuilderJersey2().build();
   private static final String RSA_PRIVATE_KEY = RsaTestHelper.generatePrivateKeyAsString();
@@ -60,7 +58,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
     assertNotNull(botAuth);
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -70,7 +68,7 @@ class AuthenticatorFactoryTest {
   void testGetBotAuthenticatorWithValidPrivateKeyInClasspath() throws AuthInitializationException {
     final BdkConfig config = createRsaConfig(() -> "classpath:/keys/private-key.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
     assertNotNull(botAuth);
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -80,7 +78,7 @@ class AuthenticatorFactoryTest {
   void testGetBotAuthenticatorWithPrivateKeyNotFoundInClasspath() {
     final BdkConfig config = createRsaConfig(() -> "classpath:/keys/notfound.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, () -> factory.getBotAuthenticator());
   }
 
@@ -90,7 +88,7 @@ class AuthenticatorFactoryTest {
     final BdkConfig config = new BdkConfig();
     config.getBot().getPrivateKey().setContent(RSA_PRIVATE_KEY.getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
 
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -102,7 +100,7 @@ class AuthenticatorFactoryTest {
     final BdkConfig config = new BdkConfig();
     config.getBot().getPrivateKey().setContent("invalid-key".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -116,7 +114,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -129,7 +127,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -138,7 +136,7 @@ class AuthenticatorFactoryTest {
 
     final BdkConfig config = createRsaConfig(() -> tempDir.resolve(UUID.randomUUID().toString() + "-privateKey.pem").toAbsolutePath().toString());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -151,7 +149,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final OboAuthenticator oboAuth = factory.getOboAuthenticator();
     assertNotNull(oboAuth);
     assertEquals(OboAuthenticatorRsaImpl.class, oboAuth.getClass());
@@ -163,7 +161,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/file.p12");
     config.getBot().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     BotAuthenticator botAuthenticator = factory.getBotAuthenticator();
     assertNotNull(botAuthenticator);
     assertEquals(BotAuthenticatorCertImpl.class, botAuthenticator.getClass());
@@ -175,7 +173,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     OboAuthenticator oboAuthenticator = factory.getOboAuthenticator();
     assertNotNull(oboAuthenticator);
     assertEquals(OboAuthenticatorCertImpl.class, oboAuthenticator.getClass());
@@ -190,7 +188,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final ExtensionAppAuthenticator botAuth = factory.getExtensionAppAuthenticator();
 
     assertEquals(ExtensionAppAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -202,7 +200,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final ExtensionAppAuthenticator extAppAuthenticator = factory.getExtensionAppAuthenticator();
 
     assertEquals(ExtensionAppAuthenticatorCertImpl.class, extAppAuthenticator.getClass());
@@ -217,7 +215,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
     config.getApp().getCertificate().setContent("certificate-content".getBytes());
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -231,7 +229,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getPrivateKey().setContent("privatekey-content".getBytes());
     config.getApp().getPrivateKey().setPath("/path/to/cert/privatekey.pem");
     config.getApp().getPrivateKey().setContent("privatekey-content".getBytes());
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -247,7 +245,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
   }
@@ -257,7 +255,7 @@ class AuthenticatorFactoryTest {
 
     final BdkConfig config = createRsaConfig(() -> tempDir.resolve(UUID.randomUUID().toString() + "-privateKey.pem").toAbsolutePath().toString());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
   }
@@ -267,7 +265,7 @@ class AuthenticatorFactoryTest {
 
     final BdkConfig config = new BdkConfig();
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getOboAuthenticator);
@@ -285,7 +283,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/file.p12");
     config.getBot().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -301,7 +299,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getOboAuthenticator);
@@ -319,7 +317,7 @@ class AuthenticatorFactoryTest {
     config.getBot().setPrivateKey(privateKey);
     config.getApp().setPrivateKey(privateKey);
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -333,7 +331,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/cert.pem");
     config.getBot().setCertificatePath("/path/to/cert/cert.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -345,7 +343,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/cert.pem");
     config.getBot().getCertificate().setContent("certificate-content".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -357,7 +355,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getPrivateKey().setPath("/path/to/cert/privatekey.pem");
     config.getBot().getPrivateKey().setContent("privatekey-content".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
