+++ symphony-bdk-spring/symphony-bdk-core-spring-boot-starter/src/main/java/com/symphony/bdk/spring/config/BdkCoreConfig.java
@@ -1,6 +1,7 @@
 package com.symphony.bdk.spring.config;
 
 import com.symphony.bdk.core.auth.AuthSession;
+import com.symphony.bdk.core.auth.impl.AuthenticatorFactoryImpl;
 import com.symphony.bdk.core.auth.AuthenticatorFactory;
 import com.symphony.bdk.core.auth.exception.AuthInitializationException;
 import com.symphony.bdk.core.auth.exception.AuthUnauthorizedException;
@@ -70,7 +71,7 @@ public class BdkCoreConfig {
   @Bean
   @ConditionalOnMissingBean
   public AuthenticatorFactory authenticatorFactory(SymphonyBdkCoreProperties properties, ApiClientFactory apiClientFactory) {
-    return new AuthenticatorFactory(properties, apiClientFactory);
+    return new AuthenticatorFactoryImpl(properties, apiClientFactory);
   }
 
   @Bean
